Refactored scene-building code and implemented the transformer scene.
