+++ faraday/src/js/views/scene/bar-magnet.js
@@ -6,10 +6,6 @@ define(function(require) {
 
     var FaradaySceneView  = require('views/scene');
 
-    var BFieldInsideView = require('views/bfield/inside');
-
-    var Assets = require('assets');
-
     // Constants
     var Constants = require('constants');
 
@@ -32,29 +28,9 @@ define(function(require) {
             this.initInsideBField();
         },
 
-        initInsideBField: function() {
-            this.bFieldInsideView = new BFieldInsideView({
-                mvt: this.mvt,
-                magnetModel: this.simulation.barMagnet,
-                needleWidth: Constants.GRID_NEEDLE_WIDTH
-            });
-
-            this.middleLayer.addChild(this.bFieldInsideView.displayObject);
-
-            this.bFieldInsideView.hide();
-        },
-
         _update: function(time, deltaTime, paused, timeScale) {
             FaradaySceneView.prototype._update.apply(this, arguments);
 
-        },
-
-        showInsideBarMagnet: function() {
-            this.bFieldInsideView.show();
-        },
-
-        hideInsideBarMagnet: function() {
-            this.bFieldInsideView.hide();
         }
 
     });
