+++ faraday/src/js/views/scene/pickup-coil.js
@@ -6,11 +6,6 @@ define(function(require) {
 
     var FaradaySceneView  = require('views/scene');
 
-    var BFieldInsideView = require('views/bfield/inside');
-    var PickupCoilView   = require('views/pickup-coil');
-
-    var Assets = require('assets');
-
     // Constants
     var Constants = require('constants');
 
@@ -36,49 +31,10 @@ define(function(require) {
             this.hideCompass();
         },
 
-        initInsideBField: function() {
-            this.bFieldInsideView = new BFieldInsideView({
-                mvt: this.mvt,
-                magnetModel: this.simulation.barMagnet,
-                needleWidth: Constants.GRID_NEEDLE_WIDTH
-            });
-
-            this.middleLayer.addChild(this.bFieldInsideView.displayObject);
-
-            this.bFieldInsideView.hide();
-        },
-
-        initPickupCoil: function() {
-            this.pickupCoilView = new PickupCoilView({
-                mvt: this.mvt,
-                model: this.simulation.pickupCoil,
-                simulation: this.simulation
-            });
-
-            this.bottomLayer.addChild(this.pickupCoilView.backgroundLayer);
-            this.topLayer.addChild(this.pickupCoilView.foregroundLayer);
-        },
-
         _update: function(time, deltaTime, paused, timeScale) {
             FaradaySceneView.prototype._update.apply(this, arguments);
 
             this.pickupCoilView.update(time, deltaTime, paused);
-        },
-
-        showInsideBarMagnet: function() {
-            this.bFieldInsideView.show();
-        },
-
-        hideInsideBarMagnet: function() {
-            this.bFieldInsideView.hide();
-        },
-
-        showPickupCoilElectrons: function() {
-            this.pickupCoilView.showElectrons();
-        },
-
-        hidePickupCoilElectrons: function() {
-            this.pickupCoilView.hideElectrons();
         }
 
     });
