+++ faraday/src/templates/electromagnet.html
@@ -27,7 +27,7 @@
             <label for="show-field-check-<%= name %>">Show Field</label>
         </li>
         <li>
-            <input type="checkbox" class="show-compass-check" id="show-compass-check-<%= name %>" checked>
+            <input type="checkbox" class="show-compass-check" id="show-compass-check-<%= name %>" <% if (!hideCompass) { %>checked<% } %>>
             <label for="show-compass-check-<%= name %>">Show Compass</label>
         </li>
         <li>
